==> notebooks/0_download_google_seed.ipynb <==
# Loading see data from BigQuery

## Your Cloud Project ID
You'll need a Google Cloud project with BigQuery enabled (it's enabled by default) for this notebook and associated code to work. Put your project ID below. Go to [cloud.google.com](http://cloud.google.com) to create one if you don't already have an account. You can create the Cloud account for free and won't be auto-billed. Then copy your Project ID and paste it below into `bq_project`.

bq_project = 'patent-embeddings'
#setup authentication with service account instead of user account
%env GOOGLE_APPLICATION_CREDENTIALS=C:\Users\tskripnikova\Documents\Patent embeddings-606d19a2d0cf.json
    
%load_ext autoreload
%autoreload 2

## Basic Configuration

import tensorflow as tf
import pandas as pd
import os

#seed_name = 'hair_dryer'
#seed_name = 'video_codec'
#seed_name = "contact_lens"
#seed_name = "contact_lens_us_c"
seed_name = "3d_printer"

seed_file = 'seeds/'+ seed_name + '.seed.csv'

src_dir = "."

patent_dataset = 'patents-public-data:patents.publications_latest'
num_anti_seed_patents = 15000
if bq_project == '':
    raise Exception('You must enter a bq_project above for this code to run.')

## Patent Landscape Expansion

This section of the notebook creates an instance of the `PatentLandscapeExpander`, which accesses a BigQuery table of patent data to do the expansion of a provided seed set and produces each expansion level as well as the final training dataset as a Pandas dataframe.

import fiz_lernmodule.expansion

expander = fiz_lernmodule.expansion.PatentLandscapeExpander(
    seed_file,
    seed_name,
    bq_project=bq_project,
    patent_dataset=patent_dataset,
    num_antiseed=num_anti_seed_patents,
    us_only=True,
    prepare_training=False)


This does the actual expansion and displays the head of the final training data dataframe.

%%time

training_data_full_df, seed_patents_df, l1_patents_df, l2_patents_df, anti_seed_patents = \
    expander.load_from_disk_or_do_expansion()

training_data_full_df.head()

### Show some stats about the landscape training data

print('Seed/Positive examples:')
print(training_data_full_df[training_data_full_df.ExpansionLevel == 'Seed'].count())

print('\n\nAnti-Seed/Negative examples:')
print(training_data_full_df[training_data_full_df.ExpansionLevel == 'AntiSeed'].count())
