==> notebooks/2_terms_attributes_parsing.ipynb <==
# Extraction of relevant terms and parsing of list attributes

#TODO: remove after development 
%load_ext autoreload
%autoreload 2

%matplotlib inline

import pandas as pd
import numpy as np
import seaborn as sns

pd.set_option('display.max_columns', 5400)
pd.set_option('max_colwidth', 100)

## Loading data

import pickle

#seed_name = 'hair_dryer'
#seed_name = 'video_codec'
#seed_name = 'diesel'
seed_name = "contact_lens"
#seed_name = "contact_lens_us_c"
#seed_name = "3d_printer"

src_dir = "."

with open(src_dir + "/data/" + seed_name + "/cleaned_text.pkl", 'rb') as infile:
    df = pickle.load(infile)

print(df.shape)
df.head()

## Extracting useful attributes

from datetime import datetime

def parse_date(date):
    date = str(date)
    try:
        datetime_object = datetime.strptime(date, '%Y%m%d')
    except ValueError:
        datetime_object = datetime.strptime("20191231", '%Y%m%d')
    return  datetime_object.year, datetime_object.month, datetime_object.day

df['year'], df['month'], df['day'] = zip(*df['priority_date'].map(parse_date))

def split_classes(s):
    no_slashes = s.replace('/', '.')
    classes = no_slashes.split(',')
    classes_with_slashes = [c.replace('.', '/') for c in classes]
    #classes_valid = [c for c in classes if len(c) <= 9]
    return sorted(classes_with_slashes)

df['cpc_classes'] = df['cpcs'].map(split_classes)

df['ipc_classes'] = df['ipcs'].map(split_classes)

def parse_ipcs(ipcs_list):
    first_letters = set()
    three_letters = set()
    four_letters = set()
    before_slash = set()
    for ipc in ipcs_list:
        if len(ipc) > 0:
            first_letters.add(ipc[0])
            three_letters.add(ipc[0:3])
            four_letters.add(ipc[0:4])
            s = ipc[0:ipc.find('/')]
            if len(s) >= 5 and len(s) <=6:
                before_slash.add(s)
    
    return sorted(list(first_letters)), sorted(list(three_letters)), sorted(list(four_letters)), sorted(list(before_slash))

df['first_letters'], df['three_letters'], df['four_letters'], df['before_slash'] = zip(*df['ipc_classes'].map(parse_ipcs))
df.head()

df['references'] = df['refs'].map(lambda r: list(filter(None, r.split(','))))
df.head()

def find_backwards_references(publication_number):
    mask = df.references.apply(lambda x: publication_number in x)
    backward_references = df[mask]  
    result = pd.unique(backward_references["publication_number"])
    return result

df['referenced_by'] = df['publication_number'].map(lambda n: find_backwards_references(n))
df.head()

from unidecode import unidecode

df['raw_assignees'] = df['assignees_harmonized'].map(lambda r: unidecode(r).split(','))

df.drop(columns = ['assignees_harmonized', 'refs', 'cpcs', 'ipcs', 'ExpansionLevel'], inplace = True)

df.head()

### Merge assignees with similar names

from fuzzywuzzy import fuzz
from fuzzywuzzy import process

def dedupe_fixed(contains_dupes, threshold=70, scorer=fuzz.token_set_ratio):
    """This convenience function takes a list of strings containing duplicates and uses fuzzy matching to identify
    and remove duplicates. Specifically, it uses the process.extract to identify duplicates that
    score greater than a user defined threshold. Then, it looks for the longest item in the duplicate list
    since we assume this item contains the most entity information and returns that. It breaks string
    length ties on an alphabetical sort.
    Note: as the threshold DECREASES the number of duplicates that are found INCREASES. This means that the
        returned deduplicated list will likely be shorter. Raise the threshold for fuzzy_dedupe to be less
        sensitive.
    Args:
        contains_dupes: A list of strings that we would like to dedupe.
        threshold: the numerical value (0,100) point at which we expect to find duplicates.
            Defaults to 70 out of 100
        scorer: Optional function for scoring matches between the query and
            an individual processed choice. This should be a function
            of the form f(query, choice) -> int.
            By default, fuzz.token_set_ratio() is used and expects both query and
            choice to be strings.
    Returns:
        A deduplicated list. For example:
            In: contains_dupes = ['Frodo Baggin', 'Frodo Baggins', 'F. Baggins', 'Samwise G.', 'Gandalf', 'Bilbo Baggins']
            In: fuzzy_dedupe(contains_dupes)
            Out: ['Frodo Baggins', 'Samwise G.', 'Bilbo Baggins', 'Gandalf']
        """

    extractor = []
    processed = []

    # iterate over items in *contains_dupes*
    for item in contains_dupes:
        if item not in processed and item not in extractor:
            # return all duplicate matches found
            matches = process.extract(item, contains_dupes, limit=None, scorer=scorer)
            # filter matches based on the threshold
            filtered = [x for x in matches if x[1] >= threshold]
            # if there is only 1 item in *filtered*, no duplicates were found so append to *extracted*
            if len(filtered) == 1:
                extractor.append(filtered[0][0])

            else:
                # alpha sort
                filtered = sorted(filtered, key=lambda x: x[0])
                # length sort
                filter_sort = sorted(filtered, key=lambda x: len(x[0]))
                #filter_sort = sorted(filtered, key=lambda x: len(x[0]), reverse=True)
                # take first item as our 'canonical example'
                extractor.append(filter_sort[0][0])
                
            processed.append([i[0] for i in filtered])

    # uniquify *extractor* list
    
    set_extractor = set(extractor)
    extractor = list(set_extractor)

    # check that extractor differs from contain_dupes (e.g. duplicates were found)
    # if not, then return the original list
    if len(extractor) == len(contains_dupes):
        return contains_dupes
    else:
        return extractor

assignees = [item for sublist in df['raw_assignees'].values for item in sublist]
assignees_set = set(assignees)
print(len(assignees_set))
#assignees_set

threshold = 88 #hair dryer

#scorer = fuzz.token_sort_ratio
scorer = fuzz.token_set_ratio

deduplicated = sorted(list(dedupe_fixed(assignees_set, threshold=threshold, scorer=scorer)))
print(len(deduplicated))

deduplicated_dictionary = {}
for d in deduplicated:
    matches = process.extractBests(d, assignees_set, score_cutoff=threshold, scorer=scorer)
    deduplicated_dictionary[d] = matches
    
#for k, v in deduplicated_dictionary.items():
#    if len(v) > 1:
#        print(k, v)

%%time

def disambiguate_assignees(assignee_list):
    result = []
    for assignee in assignee_list:
        match = process.extractOne(assignee, deduplicated)
        result.append(match[0])
    return list(set(result))

df['assignees'] = df['raw_assignees'].map(lambda x: disambiguate_assignees(x))
df.head()

## Preprocessing of tokens to labels with TF-IDF

### Text Representation
We will use <a href="https://scikit-learn.org/stable/modules/generated/sklearn.feature_extraction.text.TfidfVectorizer.html">sklearn.feature_extraction.text.TfidfVectorizer</a> to calculate a tf-idf vector for each document.<br>
Therefore, we need to define the following model parameters:

1. **sublinear_tf:** is set to True to use a logarithmic form for frequency.
2. **min_df:**  is the minimum number of documents a word must occur in (our case: more than 10).
3. **max_df:** is the maximum number of documents a word is allowed to appear in (our case: less than 30% of the documents)
4. **norm:** is set to l2, to ensure all our feature vectors have a euclidian norm of 1.
5. **ngram_range:**  is set to (1, 2) to indicate that we want to consider unigrams and bigrams.

df["token_string"] = df["tokens"].apply(lambda x: " ".join(x))

from sklearn.feature_extraction.text import TfidfVectorizer

vectorizer = TfidfVectorizer(sublinear_tf=True, min_df=10, max_df=0.2, norm='l2', encoding='latin-1', smooth_idf=True, use_idf=True, ngram_range=(1, 2))
term_doc_matrix = vectorizer.fit_transform(df["token_string"])

feature_array = np.array(vectorizer.get_feature_names())

def get_sorted_terms(x, vectorizer):
    transformed = vectorizer.transform([x])
    tfidf_sorting = np.argsort(transformed.toarray()).flatten()[::-1]
    sorted_terms = feature_array[tfidf_sorting]
    return sorted_terms

df["terms"] = df["token_string"].apply(lambda x: get_sorted_terms(x, vectorizer))

df.drop(columns=["token_string", 'text_prep', 'cpc_classes'], inplace=True)
df.head()

with open(src_dir + "/data/" + seed_name + "/terms_attributes.pkl", 'wb') as outfile:
    pickle.dump(df, outfile)
