==> notebooks/3_hierarchy.ipynb <==
# Creating different kinds of hierarchies for sunburst graph

## Loading data

%load_ext autoreload
%autoreload 2

%matplotlib inline

import pandas as pd
import numpy as np

pd.set_option('display.max_columns', 5400)
#pd.set_option('display.max_rows', 400)

import pickle 

#seed_name = 'hair_dryer'
#seed_name = 'video_codec'
#seed_name = 'diesel'
seed_name = "contact_lens"
#seed_name = "contact_lens_us_c"
#seed_name = "3d_printer"

src_dir = "../demo/data/"

with open("./data/" + seed_name + '/terms_attributes.pkl', 'rb') as infile:
    df = pickle.load(infile)
df.head()

## Computing hierarchies with different levels

### Helper functions

def unite_arrays(series):
    l = [item for sublist in series.values for item in sublist]
    myset = set(l)
    l = list(myset)
    return l

def get_unique_values(df, col, t, starts_with = ''):    
    
    if t == 'list' or t == 'code':
        values = [item for sublist in df[col].values for item in sublist]
        values_set = set(values)
        values = list(values_set)
        values_filtered = [item for item in values if item.startswith(starts_with)]
    elif t == 'value':
        values_filtered = pd.unique(df[col])    
    else:
        raise ValueError('Allowed types: list, value, code')
        
    return values_filtered

def get_subgroup(df, col, t, v):    
    
    if t == 'list' or t == 'code':
        subgroup = df.loc[df[col].apply(lambda x: v in x)] 
    elif t == 'value':
        subgroup = df.loc[df[col] == v] 
    else:
        raise ValueError('Allowed types: list, value, code')
        
    return subgroup

def create_hierarchy(df, columns, types):
    if (len(columns) != len(types)):
        raise ValueError('You must specify a type for each column')
       
    print("Creating hierarchy with columns: " + ', '.join(columns) + ' and types: ' + ', '.join(types) + '\n')
    
    hierarchy = pd.DataFrame(columns = ["id", "name", "ids", "parent", 'column'])
        
    parents = {"root": ['0_root']}
    parent_type = 'value'
    parent_col = 'unused'
    
    level_stack = []
    filtered_stack = []    
    unique_values = []
    
    print('Forward pass. Grouping\n')
    
    for i in range(len(columns)):
        t = types[i]
        col = columns[i]
        
        unique_values.append({})
        
        print(str(i) + ': Splitting by ' + t + ' attribute ' + col + '\n')   
        
        level = pd.DataFrame(columns=hierarchy.columns)
        filtered_stack_level = {}
        
        #print("Parents:\n")
        #print(parents)    
            
        for parent_key, parent_ids in parents.items():
            for parent_id in parent_ids:
                if parent_key == 'root':
                    group = df
                else:
                    group = filtered_stack[i-1][parent_id] # get_by_parent(df, parent_col, p, parent_type)                          

                #print("Grouping within parent: " + parent_id + " in parent column " + parent_col + " with type: " + parent_type)

                if t == 'code' and parent_type == 'code':
                    values = get_unique_values(group, col, t, parent_key)
                else:
                    values = get_unique_values(group, col, t)

                #print('Unique values: '+ str(len(values)) + '. Some values: ' + ', '.join(values[:10]) + '\n')

                subgroup_values = {}

                for v in values:
                    if v in unique_values[i]:
                        modified_value = str(len(unique_values[i][v])) + "_" + v                   
                        unique_values[i][v].append(modified_value)
                    else:
                        modified_value = "0_" + v
                        unique_values[i][v] = [modified_value]
                    subgroup_values[v] = modified_value

                #print("\nSubgroup values:\n")
                #print(subgroup_values)

                for key, value in subgroup_values.items(): #there should be only one value per key
                                        
                    subgroup = get_subgroup(group, col, t, key)
                    filtered_stack_level[value] = subgroup

                    ids = subgroup["publication_number"].values;
                    level = level.append(pd.DataFrame([[value, key, ids, parent_id, col]], columns = hierarchy.columns))
                
        parents = unique_values[i]
        parent_type = t
        parent_col = col
        
        #print("\nUnique values level " + str(i) + ":\n")
        #print(unique_values[i])

        #print("Filtered stack level:\n")
        #print(filtered_stack_level)
        
        level_stack.append(level)
        filtered_stack.append(filtered_stack_level)
        
        
    print('Backward pass. Aggregating the subgroup counts\n')
    
    for i in reversed(range(len(columns))):
        
        #print('i: '+ str(i))
        
        level = level_stack[i]
        
        #print('Current level:')
        #print(level.head())
        
        level_grouped = level.groupby(['parent']).agg({"ids": unite_arrays})
                
        for index, row in level_grouped.iterrows():
            parent_name = row.name
            ids = row["ids"]

            if i == 0:
                print("parent of " + parent_name + " is null")
                
                d = {"id": ["0_root"], "name": ["root"], "ids": [ids], "parent": [""], 'column':[i]}
                previous_level = pd.DataFrame(data = d)
                print("Appended root node: ")
                print(previous_level)
                
                hierarchy = hierarchy.append(previous_level)
                                        
        hierarchy = hierarchy.append(level)
               
    return hierarchy

### Dividing by assignee

hierarchy = create_hierarchy(df, ['assignees'], ['list'])

print(hierarchy.head())
hierarchy.tail()

hierarchy.to_json(src_dir + 'hierarchy/' + seed_name + "_assignee.json", orient="records")

### Dividing by country

hierarchy = create_hierarchy(df, ['country_code'], ['value'])

print(hierarchy.head())
hierarchy.tail()

hierarchy.to_json(src_dir + 'hierarchy/' + seed_name + "_country.json", orient="records")

### Dividing by all levels of IPC hierarchy: A, A21, A21B, A21B 11/04

hierarchy = create_hierarchy(df, ['first_letters', 'three_letters', 'four_letters', 'before_slash', 'ipc_classes'], ['code', 'code', 'code', 'code','code'])

print(hierarchy.head())
hierarchy.tail()

hierarchy.to_json(src_dir + 'hierarchy/' + seed_name + "_ipc_classes.json", orient="records")

### Dividing by assignee, then by country

hierarchy = create_hierarchy(df, ['assignees', 'country_code'], ['list', 'value'])

print(hierarchy.head())
hierarchy.tail()

hierarchy.to_json(src_dir + 'hierarchy/' + seed_name + "_assignee_country.json", orient="records")

### Dividing by assignee and then by upper levels of IPC hierarchy: A, A21, A21N 

hierarchy = create_hierarchy(df, ['assignees', 'first_letters', 'three_letters', 'four_letters' ], ['list', 'code', 'code', 'code'])

print(hierarchy.head())
hierarchy.tail()

hierarchy.to_json(src_dir + 'hierarchy/' + seed_name + "_assignee_ipc_classes.json", orient="records")

### Dividing by country, then by assignee

hierarchy = create_hierarchy(df, ['country_code', 'assignees'], ['value', 'list'])

print(hierarchy.head())
hierarchy.tail()

hierarchy.to_json(src_dir + 'hierarchy/' + seed_name + "_country_assignee.json", orient="records")

### Dividing by country, then by all levels of IPC hierarchy: A, A21, A21B, A21B 11/04

hierarchy = create_hierarchy(df, ['country_code', 'first_letters', 'three_letters', 'four_letters', 'before_slash', 'ipc_classes'], 
                                 ['value', 'code', 'code', 'code', 'code','code'])

print(hierarchy.head())
hierarchy.tail()

hierarchy.to_json(src_dir + 'hierarchy/' + seed_name + "_country_ipc_classes.json", orient="records")

### Dividing by upper levels of IPC hierarchy: A, A21, A21B, A21B 11 and then by assignee

hierarchy = create_hierarchy(df, ['first_letters', 'three_letters', 'four_letters', 'before_slash', 'assignees'], ['code', 'code', 'code', 'code', 'list'])

print(hierarchy.head())
hierarchy.tail()

hierarchy.to_json(src_dir + 'hierarchy/' + seed_name + "_ipc_classes_assignee.json", orient="records")

### Dividing by upper levels of IPC hierarchy: A, A21, A21B, A21B 11 and then by country

hierarchy = create_hierarchy(df, ['first_letters', 'three_letters', 'four_letters', 'before_slash', 'country_code'], ['code', 'code', 'code', 'code', 'value'])

print(hierarchy.head())
hierarchy.tail()

hierarchy.to_json(src_dir + 'hierarchy/' + seed_name + "_ipc_classes_country.json", orient="records")
